abstract out _refresh createOutput
